requirements.txt
streamlit
pandas
plotly
google-generativeai>=0.3.2
